Add optional version to test migrations
